separated the ContactForm component from the Contact route component
